requirements.txt
folium~=0.12.1
pandas~=1.1.4
numpy~=1.19.4
urllib3~=1.26.2
mysql-connector-python~=8.0.22
matplotlib~=3.3.3
scikit-learn~=0.23.2
scipy~=1.5.3
kneed~=0.7.0
holoviews~=1.14.1